install progress

+++ src/android/app/src/main/java/org/citra/emu/ui/EmulationActivity.java
@@ -460,7 +460,7 @@ public final class EmulationActivity extends AppCompatActivity {
         }
     }
 
-    public void updateProgress(String name, int written, int total) {
+    public void updateProgress(String name, long written, long total) {
         mEmulationFragment.updateProgress(name, written, total);
     }
